Fix some schema drop issues in the structure setup
I'm not entirely sure why, but locally sqlite3 had no issues dropping
tables in the order presented, but D1 gets super cranky because the
order was dropping a couple of tables at the wrong time, which was
breaking referential integrity.

It is mildly concerning that this can even happen, but, fixed now.