Integrate the UI code with this repository
This brings in the UI as static assets and alters various config files
so as to get the UI up and running inside of this project, which ties
everything together.

There is still some nascent issue where the SPA routing does not fall
back to the index page for requests that are not worker routes or
static files.

To get around this for the moment, we have a wildcard route to do it
for us, just so we're not blocked. This needs to be resolved in a better
way though.